G18-51: Create Account Models - Account Manager
I have added the Account model (now called Transaction as to not be mistaken with the existing Account model), which allows the Account Manager to view all existing transactions.
I have now split this model into two (Transaction and Booking), in order to simplify the transactions page for the the Account Manager, by seperating the individual tickets from the overall transaction itself (especially useful when considering block bookings with a lot of tickets in one transaction).